backend/pugorugh/views.py
from django.contrib.auth import get_user_model

from rest_framework import permissions
from rest_framework import viewsets
from rest_framework.generics import CreateAPIView

from . import serializers
from . import models


class UserRegisterView(CreateAPIView):
    permission_classes = (permissions.AllowAny,)
    model = get_user_model()
    serializer_class = serializers.UserSerializer


class DogViewSet(viewsets.ModelViewSet):
    queryset = models.Dog.objects.all()
    serializer_class = serializers.DogSerializer
