requirements.txt
certifi==2018.4.16
Django==1.11
djangorestframework==3.4.1
pipenv==11.10.0
pytz==2018.5
virtualenv==15.2.0
virtualenv-clone==0.3.0
